+++ R/ModelSEIRMixing.R
@@ -1,7 +1,6 @@
 #' Susceptible Exposed Infected Removed model (SEIR) with mixing
 #' @param name String. Name of the virus
 #' @param prevalence Double. Initial proportion of individuals with the virus.
-#' @param contact_rate Numeric scalar. Average number of contacts per step.
 #' @param transmission_rate Numeric scalar between 0 and 1. Probability of
 #' transmission.
 #' @param incubation_days Numeric scalar. Average number of days in the
@@ -38,11 +37,11 @@
 #' e3 <- entity("Population 3", 3e3, as_proportion = FALSE)
 #'
 #' # Row-stochastic matrix (rowsums 1)
-#' cmatrix <- c(
+#' cmatrix <- (c(
 #'   c(0.9, 0.05, 0.05),
 #'   c(0.1, 0.8, 0.1),
 #'   c(0.1, 0.2, 0.7)
-#' ) |> matrix(byrow = TRUE, nrow = 3)
+#' ) * 20) |> matrix(byrow = TRUE, nrow = 3)
 #'
 #' N <- 9e3
 #'
@@ -50,7 +49,6 @@
 #'   name              = "Flu",
 #'   n                 = N,
 #'   prevalence        = 1 / N,
-#'   contact_rate      = 20,
 #'   transmission_rate = 0.1,
 #'   recovery_rate     = 1 / 7,
 #'   incubation_days   = 7,
@@ -73,7 +71,6 @@ ModelSEIRMixing <- function(
   name,
   n,
   prevalence,
-  contact_rate,
   transmission_rate,
   incubation_days,
   recovery_rate,
@@ -83,7 +80,6 @@ ModelSEIRMixing <- function(
   stopifnot_string(name)
   stopifnot_int(n)
   stopifnot_double(prevalence)
-  stopifnot_double(contact_rate)
   stopifnot_double(transmission_rate)
   stopifnot_double(incubation_days)
   stopifnot_double(recovery_rate)
@@ -91,7 +87,7 @@ ModelSEIRMixing <- function(
 
   structure(
     ModelSEIRMixing_cpp(
-      name, n, prevalence, contact_rate,
+      name, n, prevalence,
       transmission_rate, incubation_days,
       recovery_rate, as.vector(contact_matrix)
     ),