+++ R/cpp11.R
@@ -200,8 +200,8 @@ ModelSIRMixing_cpp <- function(name, n, prevalence, contact_rate, transmission_r
   .Call(`_epiworldR_ModelSIRMixing_cpp`, name, n, prevalence, contact_rate, transmission_rate, recovery_rate, contact_matrix)
 }
 
-ModelSEIRMixing_cpp <- function(name, n, prevalence, contact_rate, transmission_rate, incubation_days, recovery_rate, contact_matrix) {
-  .Call(`_epiworldR_ModelSEIRMixing_cpp`, name, n, prevalence, contact_rate, transmission_rate, incubation_days, recovery_rate, contact_matrix)
+ModelSEIRMixing_cpp <- function(name, n, prevalence, transmission_rate, incubation_days, recovery_rate, contact_matrix) {
+  .Call(`_epiworldR_ModelSEIRMixing_cpp`, name, n, prevalence, transmission_rate, incubation_days, recovery_rate, contact_matrix)
 }
 
 ModelSEIRMixingQuarantine_cpp <- function(name, n, prevalence, contact_rate, transmission_rate, incubation_days, recovery_rate, contact_matrix, hospitalization_rate, hospitalization_period, days_undetected, quarantine_period, quarantine_willingness, isolation_willingness, isolation_period, contact_tracing_success_rate, contact_tracing_days_prior) {